Add files via upload
It was my First Year Computer Engineering Project :- A case study on Topic- Sofia(Humanoid Robot).
I was the leader of this project group. It was a  group of Four memeber each memeber is from diffrent engineering stream(Computer Engineering, Mechanical Engineering, Electronics and Telecommunication and Civil Engineering). Our motive in this project is that to highlight contribution of diffrent streams of Engineering rather than only Computer Engineering in the development of "SOFIA(Humanoid Robot).