Add files via upload
It was a Second Year Computer Engineering Project Report on Movies Recommender System using Machine Learning. I was the leader of this group and this project has two members. The project was done successfully under the guidance of "H.O,D(Computer Engineering) Dr. V.S Wadne". As Machine Learning is a huge concept and also evolving domain at present , so we focused at preparing simple Machine Learning model. We learned about machine learning and basic concepts about it. Machine Learning only require math and a good command on Programming language Python or R-language. We use Python as it is simple and has good community support. For learning and understanding ML algorithm we use "Kaggle Platform". This platform is good and free for beginner.